fixes, formatting, integrate lib

+++ contracts/contracts/Ed25519_mock.sol
@@ -1,8 +1,7 @@
 pragma solidity ^0.8.9;
 
 contract Ed25519 {
-    function scalarMultBase(uint s) public view returns (uint, uint) {
+    function scalarMultBase(uint256 s) public view returns (uint256, uint256) {
         return (s, s);
     }
-
 }
